% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{create_tdm}
\alias{create_tdm}
\title{Create TDM}
\usage{
create_tdm()
}
\value{
a term-document matrix
}
\description{
Creates a term-document matrix (TDM).
}
\details{
This function creates a default term-document matrix.
}
\author{
Michael David Gill
}

