% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{download_data_file}
\alias{download_data_file}
\title{Download Data File}
\usage{
download_data_file(location, data_file)
}
\arguments{
\item{location}{URL of the data file to be used formatted as a character.}

\item{data_file}{name of the data file to be used formatted as a character.}
}
\description{
Downloads a data file.
}
\details{
This function downloads a data file from a specified location URL for use in
the package "wordprediction".
}
\author{
Michael David Gill
}

