% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{katz_backoff}
\alias{katz_backoff}
\title{Katz's Back-Off}
\usage{
katz_backoff(phrase)
}
\arguments{
\item{phrase}{a character-type input to the model}
}
\value{
a character-type word predicted by the model
}
\description{
Creates a Katz's back-off model.
}
\details{
This function creates a Katz's back-off model.
}
\author{
Michael David Gill
}

