% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{preprocess_corpus}
\alias{preprocess_corpus}
\title{Preprocess Corpus}
\usage{
preprocess_corpus(corpus)
}
\arguments{
\item{corpus}{a volatile text corpus as a "tm" package object}
}
\value{
a preprocessed volatile text corpus as a "tm" package object
}
\description{
Preprocesses the corpus.
}
\details{
This function preprocesses a corpus by removing punctuation, numbers, and
English stopwords, and stripping whitespace.
}
\author{
Michael David Gill
}

