% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{read_data}
\alias{read_data}
\title{Read Data}
\usage{
read_data(language)
}
\arguments{
\item{language}{a character giving the language as IETF language tags "de" for German, "en"
for English, "fi" for Finnish, or "ru" for Russian.}
}
\value{
corpus a volatile text corpus as a "tm" package object
}
\description{
Reads data into a corpus.
}
\details{
This function reads from a data file as an object of type "corpus" from the
"tm" ppackage. It chooses the files corresponding to the language indicated
by the language parameter.
}
\author{
Michael David Gill
}

