% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{sample_data_file}
\alias{sample_data_file}
\title{Sample Data File}
\usage{
sample_data_file()
}
\description{
Samples a data file.
}
\details{
This function samples 1000 lines each from the data files via Mac OS X shell.
}
\author{
Michael David Gill
}

