% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{uncompress_data_file}
\alias{uncompress_data_file}
\title{Uncompress Data File}
\usage{
uncompress_data_file(location, data_file)
}
\arguments{
\item{location}{local directory created by unpacking the data file.}

\item{data_file}{name of the data file to be used formatted as a character.}
}
\description{
Uncompresses a data file
}
\details{
This function uncompresses a data file for use in the package
"wordprediction".
}
\author{
Michael David Gill
}

