% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{create_ngram}
\alias{create_ngram}
\title{Create n-Gram}
\usage{
create_ngram(corpus, n)
}
\arguments{
\item{corpus}{corpus to be tokenized}

\item{n}{size of n-gram}
}
\value{
a n-gram-tokenized term-document matrix
}
\description{
Creates a n-gram-tokenized term-document matrix (TDM).
}
\details{
This function creates a n-gram-tokenized term-document matrix.
}
\author{
Michael David Gill
}

