% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordprediction.R
\name{katz_backoff_model}
\alias{katz_backoff_model}
\title{Katz's Back-Off}
\usage{
katz_backoff_model(phrase)
}
\arguments{
\item{phrase}{a character-type input to the model}
}
\value{
a character-type word predicted by the model
}
\description{
Creates a Katz's back-off model.
}
\details{
This function creates a simplified Katz's back-off model that backs off to
smaller n-grams when the key is not found in the larger n-gram.  The maximum
n-gram handled is a trigram.  The word returned is the match found in the
largest n-gram where the key is found.  When the key is not found in the
unigram, the word "will" is returned.
}
\author{
Michael David Gill
}

